+++ runtime/filesystem-http/src/main/java/io/aklivity/zilla/runtime/filesystem/http/internal/HttpWatchService.java
@@ -45,11 +45,14 @@ import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
+import io.aklivity.zilla.runtime.filesystem.http.HttpFilesystemEnvironment;
+
 public final class HttpWatchService implements WatchService
 {
     private final WatchKey closeKey = new HttpWatchKey();
 
     private final Duration pollInterval;
+    private final String authorization;
     private final Collection<HttpWatchKey> watchKeys;
     private final BlockingQueue<WatchKey> pendingKeys;
     private final ScheduledExecutorService executor;
@@ -57,12 +60,13 @@ public final class HttpWatchService implements WatchService
     private volatile boolean closed;
 
     HttpWatchService(
-        HttpFileSystemConfiguration config)
+        HttpFilesystemEnvironment env)
     {
-        this.pollInterval = config.pollInterval();
+        this.pollInterval = env.pollInterval();
         this.watchKeys = new ConcurrentSkipListSet<>();
         this.pendingKeys = new LinkedBlockingQueue<>();
         this.executor = Executors.newScheduledThreadPool(2);
+        this.authorization = env.authorization();
     }
 
     @Override